Take photo with volume key
Closes #1630

+++ app/src/main/java/com/kylecorry/trail_sense/shared/camera/PhotoImportBottomSheetFragment.kt
@@ -4,6 +4,7 @@ import android.content.DialogInterface
 import android.net.Uri
 import android.os.Bundle
 import android.util.Size
+import android.view.KeyEvent
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -11,6 +12,7 @@ import androidx.camera.core.ImageCapture
 import androidx.camera.view.PreviewView
 import androidx.core.net.toUri
 import androidx.core.view.isInvisible
+import androidx.core.view.isVisible
 import com.kylecorry.andromeda.camera.ImageCaptureSettings
 import com.kylecorry.andromeda.fragments.BoundFullscreenDialogFragment
 import com.kylecorry.trail_sense.databinding.FragmentPhotoImportSheetBinding
@@ -24,8 +26,20 @@ class PhotoImportBottomSheetFragment(
     private val onCapture: (uri: Uri?) -> Unit
 ) : BoundFullscreenDialogFragment<FragmentPhotoImportSheetBinding>() {
 
+    private var isCapturing = false
+
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+
+        val volumeKeys = listOf(KeyEvent.KEYCODE_VOLUME_DOWN, KeyEvent.KEYCODE_VOLUME_UP)
+        dialog?.setOnKeyListener { _, keyCode, event ->
+            if (!volumeKeys.contains(keyCode) || event.action != KeyEvent.ACTION_DOWN){
+                return@setOnKeyListener false
+            }
+            takePhoto()
+            true
+        }
+
         binding.camera.setScaleType(PreviewView.ScaleType.FIT_CENTER)
         binding.camera.clipToOutline = true
         binding.camera.start(
@@ -41,21 +55,30 @@ class PhotoImportBottomSheetFragment(
         }
 
         binding.captureButton.setOnClickListener {
-            binding.captureButton.isInvisible = true
-            inBackground {
-                val file = FileSubsystem.getInstance(requireContext()).createTemp("jpg")
-                val success = onIO {
-                    binding.camera.capture(file)
+            takePhoto()
+        }
+    }
+
+    private fun takePhoto(){
+        if (isCapturing){
+            return
+        }
+        isCapturing = true
+        binding.captureButton.isInvisible = true
+        binding.loadingIndicator.isVisible = true
+        inBackground {
+            val file = FileSubsystem.getInstance(requireContext()).createTemp("jpg")
+            val success = onIO {
+                binding.camera.capture(file)
+            }
+            onMain {
+                if (!success) {
+                    file.delete()
+                    onCapture(null)
+                } else {
+                    onCapture(file.toUri())
                 }
-                onMain {
-                    if (!success) {
-                        file.delete()
-                        onCapture(null)
-                    } else {
-                        onCapture(file.toUri())
-                    }
-                    dismiss()
-                }
+                dismiss()
             }
         }
     }
